requirements.txt
fastapi==0.115.2
uvicorn==0.30.0
openai==1.53.0
python-dotenv==1.0.1
tenacity==8.2.3
pydantic==2.8.2
httpx==0.27.2
structlog==24.1.0
pytest==8.1.1